--- a/pkg/scheduler/framework/plugins/podtopologyspread/filtering_test.go
+++ b/pkg/scheduler/framework/plugins/podtopologyspread/filtering_test.go
@@ -32,7 +32,6 @@ import (
 	plugintesting "k8s.io/kubernetes/pkg/scheduler/framework/plugins/testing"
 	"k8s.io/kubernetes/pkg/scheduler/internal/cache"
 	st "k8s.io/kubernetes/pkg/scheduler/testing"
-	"k8s.io/utils/pointer"
 )
 
 var cmpOpts = []cmp.Option{
@@ -87,9 +86,9 @@ func TestPreFilterState(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone1", 0}, {"zone2", 0}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}: pointer.Int32Ptr(0),
-					{key: "zone", value: "zone2"}: pointer.Int32Ptr(0),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}: 0,
+					{key: "zone", value: "zone2"}: 0,
 				},
 			},
 		},
@@ -122,9 +121,9 @@ func TestPreFilterState(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone2", 2}, {"zone1", 3}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}: pointer.Int32Ptr(3),
-					{key: "zone", value: "zone2"}: pointer.Int32Ptr(2),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}: 3,
+					{key: "zone", value: "zone2"}: 2,
 				},
 			},
 		},
@@ -159,10 +158,10 @@ func TestPreFilterState(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone3", 0}, {"zone2", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}: pointer.Int32Ptr(3),
-					{key: "zone", value: "zone2"}: pointer.Int32Ptr(2),
-					{key: "zone", value: "zone3"}: pointer.Int32Ptr(0),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}: 3,
+					{key: "zone", value: "zone2"}: 2,
+					{key: "zone", value: "zone3"}: 0,
 				},
 			},
 		},
@@ -195,9 +194,9 @@ func TestPreFilterState(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone2", 1}, {"zone1", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}: pointer.Int32Ptr(2),
-					{key: "zone", value: "zone2"}: pointer.Int32Ptr(1),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}: 2,
+					{key: "zone", value: "zone2"}: 1,
 				},
 			},
 		},
@@ -239,13 +238,13 @@ func TestPreFilterState(t *testing.T) {
 					"zone": {{"zone1", 3}, {"zone2", 4}},
 					"node": {{"node-x", 0}, {"node-b", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(3),
-					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(4),
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(2),
-					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-x"}: pointer.Int32Ptr(0),
-					{key: "node", value: "node-y"}: pointer.Int32Ptr(4),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}:  3,
+					{key: "zone", value: "zone2"}:  4,
+					{key: "node", value: "node-a"}: 2,
+					{key: "node", value: "node-b"}: 1,
+					{key: "node", value: "node-x"}: 0,
+					{key: "node", value: "node-y"}: 4,
 				},
 			},
 		},
@@ -288,12 +287,12 @@ func TestPreFilterState(t *testing.T) {
 					"zone": {{"zone1", 3}, {"zone2", 4}},
 					"node": {{"node-b", 1}, {"node-a", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(3),
-					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(4),
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(2),
-					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-y"}: pointer.Int32Ptr(4),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}:  3,
+					{key: "zone", value: "zone2"}:  4,
+					{key: "node", value: "node-a"}: 2,
+					{key: "node", value: "node-b"}: 1,
+					{key: "node", value: "node-y"}: 4,
 				},
 			},
 		},
@@ -329,12 +328,12 @@ func TestPreFilterState(t *testing.T) {
 					"zone": {{"zone2", 0}, {"zone1", 1}},
 					"node": {{"node-a", 0}, {"node-y", 0}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(1),
-					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(0),
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(0),
-					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-y"}: pointer.Int32Ptr(0),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}:  1,
+					{key: "zone", value: "zone2"}:  0,
+					{key: "node", value: "node-a"}: 0,
+					{key: "node", value: "node-b"}: 1,
+					{key: "node", value: "node-y"}: 0,
 				},
 			},
 		},
@@ -375,12 +374,12 @@ func TestPreFilterState(t *testing.T) {
 					"zone": {{"zone1", 3}, {"zone2", 4}},
 					"node": {{"node-b", 0}, {"node-a", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(3),
-					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(4),
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-b"}: pointer.Int32Ptr(0),
-					{key: "node", value: "node-y"}: pointer.Int32Ptr(2),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}:  3,
+					{key: "zone", value: "zone2"}:  4,
+					{key: "node", value: "node-a"}: 1,
+					{key: "node", value: "node-b"}: 0,
+					{key: "node", value: "node-y"}: 2,
 				},
 			},
 		},
@@ -423,12 +422,12 @@ func TestPreFilterState(t *testing.T) {
 					"zone": {{"zone1", 3}, {"zone2", 4}},
 					"node": {{"node-b", 1}, {"node-a", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(3),
-					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(4),
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(2),
-					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-y"}: pointer.Int32Ptr(4),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}:  3,
+					{key: "zone", value: "zone2"}:  4,
+					{key: "node", value: "node-a"}: 2,
+					{key: "node", value: "node-b"}: 1,
+					{key: "node", value: "node-y"}: 4,
 				},
 			},
 		},
@@ -460,7 +459,7 @@ func TestPreFilterState(t *testing.T) {
 					"node": newCriticalPaths(),
 					"rack": newCriticalPaths(),
 				},
-				TpPairToMatchNum: make(map[topologyPair]*int32),
+				TpPairToMatchNum: make(map[topologyPair]int),
 			},
 		},
 		{
@@ -496,7 +495,7 @@ func TestPreFilterState(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": newCriticalPaths(),
 				},
-				TpPairToMatchNum: make(map[topologyPair]*int32),
+				TpPairToMatchNum: make(map[topologyPair]int),
 			},
 		},
 		{
@@ -568,9 +567,9 @@ func TestPreFilterStateAddPod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"node": {{"node-b", 0}, {"node-a", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-b"}: pointer.Int32Ptr(0),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "node", value: "node-a"}: 1,
+					{key: "node", value: "node-b"}: 0,
 				},
 			},
 		},
@@ -593,9 +592,9 @@ func TestPreFilterStateAddPod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"node": {{"node-a", 1}, {"node-b", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "node", value: "node-a"}: 1,
+					{key: "node", value: "node-b"}: 1,
 				},
 			},
 		},
@@ -618,9 +617,9 @@ func TestPreFilterStateAddPod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"node": {{"node-a", 0}, {"node-b", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(0),
-					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "node", value: "node-a"}: 0,
+					{key: "node", value: "node-b"}: 1,
 				},
 			},
 		},
@@ -643,9 +642,9 @@ func TestPreFilterStateAddPod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"node": {{"node-a", 0}, {"node-b", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(0),
-					{key: "node", value: "node-b"}: pointer.Int32Ptr(2),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "node", value: "node-a"}: 0,
+					{key: "node", value: "node-b"}: 2,
 				},
 			},
 		},
@@ -668,11 +667,11 @@ func TestPreFilterStateAddPod(t *testing.T) {
 					"zone": {{"zone2", 0}, {"zone1", 1}},
 					"node": {{"node-x", 0}, {"node-a", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(1),
-					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(0),
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-x"}: pointer.Int32Ptr(0),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}:  1,
+					{key: "zone", value: "zone2"}:  0,
+					{key: "node", value: "node-a"}: 1,
+					{key: "node", value: "node-x"}: 0,
 				},
 			},
 		},
@@ -697,11 +696,11 @@ func TestPreFilterStateAddPod(t *testing.T) {
 					"zone": {{"zone1", 1}, {"zone2", 1}},
 					"node": {{"node-a", 1}, {"node-x", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(1),
-					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(1),
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-x"}: pointer.Int32Ptr(1),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}:  1,
+					{key: "zone", value: "zone2"}:  1,
+					{key: "node", value: "node-a"}: 1,
+					{key: "node", value: "node-x"}: 1,
 				},
 			},
 		},
@@ -729,12 +728,12 @@ func TestPreFilterStateAddPod(t *testing.T) {
 					"zone": {{"zone2", 1}, {"zone1", 3}},
 					"node": {{"node-a", 1}, {"node-x", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(3),
-					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(1),
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-b"}: pointer.Int32Ptr(2),
-					{key: "node", value: "node-x"}: pointer.Int32Ptr(1),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}:  3,
+					{key: "zone", value: "zone2"}:  1,
+					{key: "node", value: "node-a"}: 1,
+					{key: "node", value: "node-b"}: 2,
+					{key: "node", value: "node-x"}: 1,
 				},
 			},
 		},
@@ -769,12 +768,12 @@ func TestPreFilterStateAddPod(t *testing.T) {
 					"zone": {{"zone2", 1}, {"zone1", 2}},
 					"node": {{"node-a", 0}, {"node-b", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(2),
-					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(1),
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(0),
-					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-x"}: pointer.Int32Ptr(2),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}:  2,
+					{key: "zone", value: "zone2"}:  1,
+					{key: "node", value: "node-a"}: 0,
+					{key: "node", value: "node-b"}: 1,
+					{key: "node", value: "node-x"}: 2,
 				},
 			},
 		},
@@ -809,12 +808,12 @@ func TestPreFilterStateAddPod(t *testing.T) {
 					"zone": {{"zone1", 1}, {"zone2", 1}},
 					"node": {{"node-a", 1}, {"node-b", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(1),
-					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(1),
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-x"}: pointer.Int32Ptr(2),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}:  1,
+					{key: "zone", value: "zone2"}:  1,
+					{key: "node", value: "node-a"}: 1,
+					{key: "node", value: "node-b"}: 1,
+					{key: "node", value: "node-x"}: 2,
 				},
 			},
 		},
@@ -889,9 +888,9 @@ func TestPreFilterStateRemovePod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone1", 1}, {"zone2", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}: pointer.Int32Ptr(1),
-					{key: "zone", value: "zone2"}: pointer.Int32Ptr(1),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}: 1,
+					{key: "zone", value: "zone2"}: 1,
 				},
 			},
 		},
@@ -919,9 +918,9 @@ func TestPreFilterStateRemovePod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone1", 1}, {"zone2", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}: pointer.Int32Ptr(1),
-					{key: "zone", value: "zone2"}: pointer.Int32Ptr(2),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}: 1,
+					{key: "zone", value: "zone2"}: 2,
 				},
 			},
 		},
@@ -950,9 +949,9 @@ func TestPreFilterStateRemovePod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone1", 2}, {"zone2", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}: pointer.Int32Ptr(2),
-					{key: "zone", value: "zone2"}: pointer.Int32Ptr(2),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}: 2,
+					{key: "zone", value: "zone2"}: 2,
 				},
 			},
 		},
@@ -981,9 +980,9 @@ func TestPreFilterStateRemovePod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone1", 2}, {"zone2", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}: pointer.Int32Ptr(2),
-					{key: "zone", value: "zone2"}: pointer.Int32Ptr(2),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}: 2,
+					{key: "zone", value: "zone2"}: 2,
 				},
 			},
 		},
@@ -1013,12 +1012,12 @@ func TestPreFilterStateRemovePod(t *testing.T) {
 					"zone": {{"zone2", 1}, {"zone1", 3}},
 					"node": {{"node-b", 1}, {"node-x", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]*int32{
-					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(3),
-					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(1),
-					{key: "node", value: "node-a"}: pointer.Int32Ptr(2),
-					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
-					{key: "node", value: "node-x"}: pointer.Int32Ptr(1),
+				TpPairToMatchNum: map[topologyPair]int{
+					{key: "zone", value: "zone1"}:  3,
+					{key: "zone", value: "zone2"}:  1,
+					{key: "node", value: "node-a"}: 2,
+					{key: "node", value: "node-b"}: 1,
+					{key: "node", value: "node-x"}: 1,
 				},
 			},
 		},