#### What type of PR is this?

/kind cleanup

#### What this PR does / why we need it:

As documented in #105750, the performance of the pod topology spread plugin degrades as the number of pods increases. The following CPU profile of the scheduler was taken during a benchmarking run with the same setup described in #105750:

<img width="1463" alt="Screen Shot 2021-12-13 at 6 19 22 PM" src="https://user-images.githubusercontent.com/5430846/150018127-270a0035-f8c6-4cea-bb17-4ba37693003d.png">

A disproportionate amount of CPU time is spent in the `calPreFilterState` function, as expected. Let's zoom in on this function:

<img width="804" alt="Screen Shot 2021-12-13 at 6 25 20 PM" src="https://user-images.githubusercontent.com/5430846/150018561-a9ad5e78-8e8f-498f-a680-30d9a94d3db6.png">

Interestingly, the cumulative CPU time spent on `mapaccess2_faststr` accounts for nearly half (9.15/19.88 = 0.46) of the CPU time of `calPreFilterState`. pprof's list command can help us locate the bottlenecks:

```
(pprof) list calPreFilterState
     7.50s     19.88s (flat, cum) 34.39% of Total

        3s     10.63s    238:		if !nodeLabelsMatchSpreadConstraints(node.Labels, constraints) {
     130ms      6.80s    257:		count := countPodsMatchSelector(nodeInfo.Pods, constraint.Selector, pod.Namespace)

(pprof) list nodeLabelsMatchSpreadConstraints
     110ms      7.63s (flat, cum) 13.20% of Total

      40ms      7.56s     62:		if _, ok := nodeLabels[c.TopologyKey]; !ok {
```

A disproportionate amount of CPU time is spent accessing node labels; these accesses are currently performed in sequence. This PR relaxes this bottleneck by parallelizing the node label accesses. Note that `calPreFilterState` now accesses each key of `TpPairToMatchNum` in sequence, which amounts to the same number of sequential map accesses as before. This is faster because the accesses to `TpPairToMatchNum` exhibit better spatial locality and are less likely to incur a memory access. Our benchmarks (methodology described in #105750) yielded the following reductions in PreFilter latencies with a single topology constraint for hostname.

```
p50: 41.0% reduction
p90: 12.0% reduction
p99: 21.0% reduction
```

```release-note
NONE
```