@Huang-Wei 2022-01-27
> LGTM overall. Could you squash the commits?

@bbarnes52 would you mind squshing the commits into one commit? Then I think it's good to be merged. BTW: as mentioned in today's sig-meeting, could you help to summarize the profiling procedure into a practical document, in https://github.com/kubernetes/community, under the folder `contributors/devel/sig-scheduling`. I believe this will be of great help to the whole community.

FYI @alculquicondor this is the PR I mentioned in today's meeting.