@alculquicondor 2022-01-27
I'd like to take a look as well if you don't mind, as I've personally tried to optimize these memory accesses myself in the past.

@bbarnes52, do you think there are learnings from here that could be applied to the PreScore/Score extension points?