@Huang-Wei 2022-02-03
Ping @bbarnes52, could you followup on https://github.com/kubernetes/kubernetes/pull/107623#issuecomment-1023527006?

> @bbarnes52 would you mind squshing the commits into one commit? Then I think it's good to be merged. BTW: as mentioned in today's sig-meeting, could you help to summarize the profiling procedure into a practical document, in https://github.com/kubernetes/community, under the folder contributors/devel/sig-scheduling. I believe this will be of great help to the whole community.