@bbarnes52 2022-02-04
> I'd like to take a look as well if you don't mind, as I've personally tried to optimize these memory accesses myself in the past.
> 
> @bbarnes52, do you think there are learnings from here that could be applied to the PreScore/Score extension points?

Good question. I took a look at parallelizing node label accesses in [initPreScoreState](https://github.com/kubernetes/kubernetes/blob/v1.24.0-alpha.2/pkg/scheduler/framework/plugins/podtopologyspread/scoring.go#L77). This is less of a bottleneck because it iterates over filtered nodes and not all nodes. The benchmarks I performed yielded statistically insignificant results, I do not believe it is worth introducing additional synchronization at this time.

Regarding capturing the benchmarking procedure in a practical document, I would suggest Russ Cox's [Profiling Go Programs](https://go.dev/blog/pprof), which describes all the methods used here and more.