@alculquicondor 2022-02-04
/lgtm

Thanks!

Do you plan to evaluate what can be done to improve Score?