@bbarnes52 2022-02-04
> Do you plan to evaluate what can be done to improve Score?

Good question. Score appears to be less of a performance bottleneck. The latencies of PreScore/Score from my benchmarks are reproduced below (all units in microseconds). I've also attached a CPU profile. This plugin's Score function is not shown because it consumes comparably little CPU.

Looking at the code, I cannot identify any non-intrusive changes that would yield performance improvements. 

```
   "PreScore": {
      "Perc50": 2212528,
      "Perc90": 3526917,
      "Perc99": 6195231
    },
    "Score": {
      "Perc50": 950042,
      "Perc90": 1478549,
      "Perc99": 1597463
    },
```


<img width="1175" alt="Screen Shot 2022-01-25 at 11 07 22 AM" src="https://user-images.githubusercontent.com/5430846/152588865-72bf6dcb-8d4d-4142-84ce-3ddf1ea607a2.png">