@alculquicondor 2022-02-04
Yeah, I squeezed the performance of Score as much as I could. But I thought this might help across the codebase #107504